backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28
typing_extensions==4.10.0
Werkzeug==3.0.1
authlib>=1.2.0
requests>=2.28.2
python-dotenv>=0.19.2
authlib>=1.0
flask_session>=0.8.0